Data.Conduino

Documentation

data Pipe i o u m a

Similar to Conduit

i: Type of input stream

o: Type of output stream

u: Type of the result of the upstream pipe (Outputted when upstream pipe finishes)

m: Underlying monad

a: Result type (Outputted when finished)

Some specializations:

A pipe is a source if i is (): it doesn't need anything to go pump out items.

If a pipe is source and a is Void, it means that it will produce forever.

A pipe is a sink if o is Void: it will never yield anything else downstream.

If a pipe is both a source and a sink, it is an effect.

Normally you can ask for input upstream with await, which returns Nothing if the pipe upstream stops producing. However, if u is Void, it means that the pipe upstream will never stop, so you can use awaitSurely to get a guaranteed answer.

(.|) :: Monad m => Pipe a b u m v -> Pipe b c v m r -> Pipe a c u m r infixr 2

runPipe :: Monad m => Pipe () Void u m a -> m a

awaitEither :: Pipe i o u m (Either i u)

await :: Pipe i o u m (Maybe i)

awaitSurely :: Pipe i o Void m i

repeatM :: Monad m => m o -> Pipe i o u m u

unfoldP :: (b -> Maybe (a, b)) -> b -> Pipe i a u m ()

unfoldPForever :: (b -> (a, b)) -> b -> Pipe i a u m r

iterateP :: (a -> a) -> a -> Pipe i a u m r

sourceList :: Foldable t => t a -> Pipe i a u m ()

awaitForever :: (i -> Pipe i o u m a) -> Pipe i o u m u

mapP :: (a -> b) -> Pipe a b u m u

mapMP :: Monad m => (a -> m b) -> Pipe a b u m u

dropP :: Int -> Pipe i o u m ()

foldrP :: (a -> b -> b) -> b -> Pipe a Void u m b

sinkList :: Pipe i Void u m [i]

newtype ZipSink i u m a

Constructors

ZipSink

Fields

getZipSink :: Pipe i Void u m a