Data.Conduino.Internal

Documentation

newtype Pipe i o u m a

Similar to Conduit

i: Type of input stream

o: Type of output stream

u: Type of the result of the upstream pipe (Outputted when upstream pipe finishes)

m: Underlying monad

a: Result type (Outputted when finished)

Some specializations:

A pipe is a source if i is (): it doesn't need anything to go pump out items. If a pipe is source and a is Void, it means that it will produce forever.

A pipe is a sink if o is Void: it will never yield anything else downstream.

If a pipe is both a source and a sink, it is an effect.

Normally you can ask for input upstream with await, which returns Nothing if the pipe upstream stops producing. However, if u is Void, it means that the pipe upstream will never stop, so you can use awaitSurely to get a guaranteed answer.

Constructors

Pipe

Fields

pipeFree :: FT (PipeF i o u) m a

data PipeF i o u a

Constructors

PAwaitF (i -> a) (u -> a)

PYieldF o a

awaitEither :: Pipe i o u m (Either i u)

yield :: o -> Pipe i o u m ()