Data.Conduino

Documentation

data Pipe i o u m a

Similar to Conduit

i: Type of input stream

o: Type of output stream

u: Type of the result of the upstream pipe (Outputted when upstream pipe finishes)

m: Underlying monad

a: Result type (Outputted when finished)

Some specializations:

A pipe is a source if i is (): it doesn't need anything to go pump out items. If a pipe is source and a is Void, it means that it will produce forever.

A pipe is a sink if o is Void: it will never yield anything else downstream.

If a pipe is both a source and a sink, it is an effect.

Normally you can ask for input upstream with await, which returns Nothing if the pipe upstream stops producing. However, if u is Void, it means that the pipe upstream will never stop, so you can use awaitSurely to get a guaranteed answer.

(.|) :: Monad m => Pipe a b u m v -> Pipe b c v m r -> Pipe a c u m r infixr 2

The main operator for chaining pipes together. pipe1 .| pipe2 will connect the output of pipe1 to the input of pipe2.

Running a pipe will draw from pipe2, and if pipe2 ever asks for input (with await or something similar), it will block until pipe1 outputs something (or signals termination).

The structure of a full pipeline usually looks like:

runPipe $ someSource
         .| somePipe
         .| someOtherPipe
         .| someSink

Where you route a source into a series of pipes, which eventually ends up at a sink. runPipe will then produce the result of that sink.

runPipe :: Monad m => Pipe () Void u m a -> m a

Run a pipe that is both a source and a sink into the effect that it represents.

Usually you wouild construct this using something like:

runPipe $ someSource
         .| somePipe
         .| someOtherPipe
         .| someSink

runPipe will produce the result of that sink.

awaitEither :: Pipe i o u m (Either u i)

Await on upstream output. Will block until it receives an i (expected input type) or a u if the upstream pipe terminates.

await :: Pipe i o u m (Maybe i)

Await input from upstream. Will block until upstream yields.

Will return Nothing if the upstream pipe finishes and terminates.

If the upstream pipe never terminates, then you can use awaitSurely to guarantee a result.

awaitSurely :: Pipe i o Void m i

Await input from upstream where the upstream pipe is guaranteed to never terminate.

awaitForever :: (i -> Pipe i o u m a) -> Pipe i o u m u

A useful utility function over repeated awaits. Will repeatedly await and then continue with the given pipe whenever the upstream pipe yields.

Can be used to implement many pipe combinators:

map f = awaitForever $ x -> yield (f x)

yield :: o -> Pipe i o u m ()

Send output downstream.

mapInput :: (i -> j) -> Pipe j o u m a -> Pipe i o u m a

(Contravariantly) map over the expected input type.

mapOutput :: (p -> o) -> Pipe i p u m a -> Pipe i o u m a

Map over the downstream output type.

If you want to map over the result type, use fmap.

mapUpRes :: (u -> v) -> Pipe i o v m a -> Pipe i o u m a

(Contravariantly) map over the upstream result type.

trimapPipe :: (i -> j) -> (p -> o) -> (u -> v) -> Pipe j p v m a -> Pipe i o u m a

Map over the input type, output type, and upstream result type.

If you want to map over the result type, use fmap.

newtype ZipSink i u m a

A newtype wrapper over a sink (Pipe i Void) that gives it an alternative Applicative and Alternative instance.

<*> will distribute input over both sinks, and output a final result once both sinks finish.

<|> will distribute input over both sinks, and output a final result as soon as one or the other finishes.

Constructors

ZipSink

Fields

getZipSink :: Pipe i Void u m a