Data.Conduino.Internal

Documentation

newtype Pipe i o u m a

Similar to Conduit

i: Type of input stream

o: Type of output stream

u: Type of the result of the upstream pipe (Outputted when upstream pipe finishes)

m: Underlying monad

a: Result type (Outputted when finished)

Some specializations:

A pipe is a source if i is (): it doesn't need anything to go pump out items. If a pipe is source and a is Void, it means that it will produce forever.

A pipe is a sink if o is Void: it will never yield anything else downstream.

If a pipe is both a source and a sink, it is an effect.

Normally you can ask for input upstream with await, which returns Nothing if the pipe upstream stops producing. However, if u is Void, it means that the pipe upstream will never stop, so you can use awaitSurely to get a guaranteed answer.

Constructors

Pipe

Fields

pipeFree :: FT (PipeF i o u) m a

data PipeF i o u a

Base functor of Pipe.

Constructors

PAwaitF (u -> a) (i -> a)

PYieldF o a

awaitEither :: Pipe i o u m (Either u i)

Await on upstream output. Will block until it receives an i (expected input type) or a u if the upstream pipe terminates.

yield :: o -> Pipe i o u m ()

Send output downstream.

trimapPipe :: (i -> j) -> (p -> o) -> (u -> v) -> Pipe j p v m a -> Pipe i o u m a

Map over the input type, output type, and upstream result type.

If you want to map over the result type, use fmap.

mapInput :: (i -> j) -> Pipe j o u m a -> Pipe i o u m a

(Contravariantly) map over the expected input type.

mapOutput :: (p -> o) -> Pipe i p u m a -> Pipe i o u m a

Map over the downstream output type.

If you want to map over the result type, use fmap.

mapUpRes :: (u -> v) -> Pipe i o v m a -> Pipe i o u m a

(Contravariantly) map over the upstream result type.

hoistPipe :: (Monad m, Monad n) => (forall x. m x -> n x) -> Pipe i o u m a -> Pipe i o u n a

Transform the underlying monad of a pipe.

type RecPipe i o u = FreeT (PipeF i o u)

A version of Pipe that uses explicit, concrete recursion instead of church-encoding like Pipe. Some functions --- especially ones that combine multiple pipes into one --- are easier to implement in this form.

toRecPipe :: Monad m => Pipe i o u m a -> RecPipe i o u m a

Convert from a Pipe to a RecPipe. While most of this library is defined in terms of Pipe, it can be easier to write certain low-level pipe combining functions in terms of RecPipe than Pipe.

fromRecPipe :: Monad m => RecPipe i o u m a -> Pipe i o u m a

Convert a RecPipe back into a Pipe.

annotatePipeF :: PipeF i o u a -> (Const String :*: PipeF i o u) a